+++ app/lib/modules/quran/data/data.dart
@@ -5,3 +5,4 @@ export 'source/juz_data.dart';
 export 'source/juz_local_data_source.dart';
 export 'source/surah_data.dart';
 export 'source/surah_local_data_source.dart';
+export 'repository/quran_repository_impl.dart';
