+++ app/lib/modules/quran/data/source/juz_local_data_source.dart
@@ -3,27 +3,8 @@ import 'package:my_quran/modules/modules.dart';
 
 @immutable
 final class JuzLocalDataSource {
-  List<JuzEntity> getJuzsFromLocal() {
-    final juzs = juzData.map((data) => _convertJuzData(JuzResponse.fromJson(data))).toList();
+  List<JuzResponse> getJuzsFromLocal() {
+    final juzs = juzData.map(JuzResponse.fromJson).toList();
     return juzs;
-  }
-
-  JuzEntity _convertJuzData(JuzResponse response) {
-    final convertedSurahs = response.surahs
-        .map(
-          (juzSurahResponse) => JuzSurahEntity(
-            name: juzSurahResponse.name,
-            arName: juzSurahResponse.arName,
-            pages: juzSurahResponse.pages,
-          ),
-        )
-        .toList();
-
-    return JuzEntity(
-      id: response.id,
-      name: response.name,
-      pages: response.pages,
-      surahs: convertedSurahs,
-    );
   }
 }
