+++ app/lib/modules/quran/data/source/surah_local_data_source.dart
@@ -3,18 +3,8 @@ import 'package:my_quran/modules/modules.dart';
 
 @immutable
 final class SurahLocalDataSource {
-  List<SurahEntity> getSurahsFromLocal() {
-    final surahs = surahData.map((data) => _convertSurahData(SurahResponse.fromJson(data))).toList();
+  List<SurahResponse> getSurahsFromLocal() {
+    final surahs = surahData.map(SurahResponse.fromJson).toList();
     return surahs;
-  }
-
-  SurahEntity _convertSurahData(SurahResponse response) {
-    return SurahEntity(
-      id: response.id,
-      name: response.name,
-      aya: response.aya,
-      pages: response.pages,
-      arabic: response.arabic,
-    );
   }
 }
